apps/exchange-ingest/requirements.txt
websocket-client==1.8.0
requests==2.32.3
psycopg2-binary==2.9.10
